refactor(card): index 기반의 파일 네이밍을 적용한다
